gpu_programming/requirements.txt
numpy==1.16.2
numba==0.43.0
Pillow==5.4.1